+++ app/scripts/controllers/kwic_controller.ts
@@ -3,7 +3,7 @@ import angular, { IController, ITimeoutService } from "angular"
 import _ from "lodash"
 import settings from "@/settings"
 import kwicProxyFactory, { type KwicProxy } from "@/backend/kwic-proxy"
-import { ApiKwic, Response, ProgressReport } from "@/backend/types"
+import { ApiKwic, ProgressReport } from "@/backend/types"
 import { QueryParams, QueryResponse } from "@/backend/types/query"
 import { RootScope } from "@/root-scope.types"
 import { LocationService } from "@/urlparams"
@@ -18,7 +18,6 @@ export type KwicCtrlScope = TabHashScope & {
     aborted?: boolean
     buildQueryOptions: (isPaging: boolean) => QueryParams
     corpusHits?: Record<string, number>
-    countCorpora?: () => number | null
     corpusOrder?: string[]
     cqp?: string
     error?: string
@@ -261,10 +260,6 @@ export class KwicCtrl implements IController {
 
         s.onexit = () => {
             s.active = false
-        }
-
-        s.countCorpora = () => {
-            return s.proxy.prevParams?.corpus ? s.proxy.prevParams.corpus.split(",").length : null
         }
     }
 }