+++ app/scripts/root-scope.types.ts
@@ -1,7 +1,6 @@
 /** @format */
 import { IDeferred, IPromise, IRootScopeService } from "angular"
-import { Settings } from "./settings/settings.types"
-import { LangLocMap, LocLangMap } from "@/i18n/types"
+import { LangLocMap } from "@/i18n/types"
 import { KorpQueryRequestOptions } from "./backend/kwic-proxy"
 import { CqpQuery } from "./cqp_parser/cqp.types"
 import { CorpusListing } from "./corpus_listing"
@@ -9,7 +8,6 @@ import { CompareResult, MapRequestResult } from "@/backend/backend"
 
 /** Extends the Angular Root Scope interface with properties used by this app. */
 export type RootScope = IRootScopeService & {
-    _settings: Settings
     activeSearch: {
         /** "word", "lemgram" or "cqp" */
         type: string
@@ -18,7 +16,6 @@ export type RootScope = IRootScopeService & {
     extendedCQP: string | null
     globalFilter: CqpQuery | null
     globalFilterDef: IDeferred<never>
-    searchtabs: any
     simpleCQP?: string
     show_modal: "about" | false
     compareTabs: CompareTab[]
@@ -29,13 +26,6 @@ export type RootScope = IRootScopeService & {
     waitForLogin: boolean
     lang: string
     loc_data: LangLocMap
-    openErrorModal: (options: {
-        content: string
-        resolvable?: boolean
-        onClose?: () => void
-        buttonText?: string
-        translations?: LocLangMap
-    }) => void
     $on: (name: "corpuschooserchange", handler: (event: any, selected: string[]) => void) => void
 }